+++ src/audio.c
@@ -1,6 +1,7 @@
 #define _GNU_SOURCE
-#include <json-c/json.h>
 #include "utils.h"
+#include <errno.h>
+#include <json-c/json.h>
 #include <libgen.h>
 #include <pthread.h>
 #include <pulse/error.h>
@@ -12,7 +13,6 @@
 #include <sys/select.h>
 #include <sys/time.h>
 #include <unistd.h>
-#include <errno.h>
 
 #define MAX_LINE_LENGTH 1024
 #define MAX_CONCURRENT_SOUNDS 10
@@ -47,10 +47,11 @@ float g_volume = 1.0f;
 float g_mouse_volume = 1.0f;
 int g_verbose = 0;
 int g_mute = 0;
+int g_keyboard_mute = 0;
+int g_mouse_mute = 0;
 pthread_t sound_threads[MAX_CONCURRENT_SOUNDS];
 volatile int thread_active[MAX_CONCURRENT_SOUNDS] = {0};
 pthread_mutex_t thread_mutex = PTHREAD_MUTEX_INITIALIZER;
-
 
 static void get_full_path(char *buffer, size_t buffer_size,
                           const char *base_dir, const char *filename) {
@@ -121,7 +122,8 @@ int load_sound_config(const char *config_path) {
     g_sound_pack.num_generic_press_files = 0;
     if (json_object_object_get_ex(root, "sound", &obj)) {
       const char *pattern = json_object_get_string(obj);
-      if (g_verbose) printf("Sound pattern: %s\n", pattern);
+      if (g_verbose)
+        printf("Sound pattern: %s\n", pattern);
       if (strstr(pattern, "%d") || strstr(pattern, "{")) {
         for (int i = 0; i <= 4; i++) {
           char temp_filename[256];
@@ -176,13 +178,14 @@ int load_sound_config(const char *config_path) {
       get_full_path(g_sound_pack.release_file,
                     sizeof(g_sound_pack.release_file), config_dir,
                     temp_release_file);
-      if (g_verbose) printf("Release sound file: %s\n", g_sound_pack.release_file);
+      if (g_verbose)
+        printf("Release sound file: %s\n", g_sound_pack.release_file);
     }
     if (json_object_object_get_ex(root, "defines", &obj)) {
       json_object_object_foreach(obj, key, val) {
         int key_code;
         int is_release = 0;
-        
+
         // Handle mouse key names first
         if (strcmp(key, "MouseLeft") == 0) {
           key_code = 272;
@@ -203,7 +206,7 @@ int load_sound_config(const char *config_path) {
             key_code = atoi(key);
           }
         }
-        
+
         if (key_code >= 0 && key_code < 512) {
           const char *filename_relative = json_object_get_string(val);
           char full_filename[MAX_LINE_LENGTH];
@@ -233,7 +236,8 @@ int load_sound_config(const char *config_path) {
       temp_sound_file[sizeof(temp_sound_file) - 1] = '\0';
       get_full_path(g_sound_pack.sound_file, sizeof(g_sound_pack.sound_file),
                     config_dir, temp_sound_file);
-      if (g_verbose) printf("Single mode sound file: %s\n", g_sound_pack.sound_file);
+      if (g_verbose)
+        printf("Single mode sound file: %s\n", g_sound_pack.sound_file);
     }
     // Try "defines" first, then "definitions" for mouse configs
     json_object *defines_obj = NULL;
@@ -241,7 +245,7 @@ int load_sound_config(const char *config_path) {
         json_object_object_get_ex(root, "definitions", &defines_obj)) {
       json_object_object_foreach(defines_obj, key, val) {
         int key_code;
-        
+
         // Handle mouse key names
         if (strcmp(key, "MouseLeft") == 0) {
           key_code = 272;
@@ -252,7 +256,7 @@ int load_sound_config(const char *config_path) {
         } else {
           key_code = atoi(key);
         }
-        
+
         if (key_code >= 0 && key_code < 512) {
           // Handle different config formats
           if (json_object_is_type(val, json_type_array)) {
@@ -264,19 +268,23 @@ int load_sound_config(const char *config_path) {
                   json_object_get_int(json_object_array_get_idx(val, 1));
             }
           } else if (json_object_is_type(val, json_type_object)) {
-            // Mouse format: {"timing": [[start1, duration1], [start2, duration2]]}
+            // Mouse format: {"timing": [[start1, duration1], [start2,
+            // duration2]]}
             json_object *timing_array;
             if (json_object_object_get_ex(val, "timing", &timing_array) &&
                 json_object_is_type(timing_array, json_type_array) &&
                 json_object_array_length(timing_array) > 0) {
               // Use the first timing entry
-              json_object *first_timing = json_object_array_get_idx(timing_array, 0);
+              json_object *first_timing =
+                  json_object_array_get_idx(timing_array, 0);
               if (json_object_is_type(first_timing, json_type_array) &&
                   json_object_array_length(first_timing) >= 2) {
                 g_sound_pack.key_mappings[key_code].start_ms =
-                    json_object_get_int(json_object_array_get_idx(first_timing, 0));
+                    json_object_get_int(
+                        json_object_array_get_idx(first_timing, 0));
                 g_sound_pack.key_mappings[key_code].duration_ms =
-                    json_object_get_int(json_object_array_get_idx(first_timing, 1));
+                    json_object_get_int(
+                        json_object_array_get_idx(first_timing, 1));
               }
             }
           }
@@ -322,16 +330,26 @@ void *play_sound_thread(void *arg) {
   int thread_id = data->thread_id;
   int is_pressed = data->is_pressed;
   const char *file_to_play = NULL;
-  
+
   // Determine if this is a mouse event (key codes 272, 273, 274)
   int is_mouse_event = (key_code == 272 || key_code == 273 || key_code == 274);
   SoundPack *sound_pack = is_mouse_event ? &g_mouse_sound_pack : &g_sound_pack;
   float volume = is_mouse_event ? g_mouse_volume : g_volume;
-  
+  int mute_state = is_mouse_event ? g_mouse_mute : g_keyboard_mute;
+
   if (g_verbose) {
     printf("Thread %d: Playing %s sound for key %d (%s)\n", thread_id,
            is_mouse_event ? "mouse" : "keyboard", key_code,
            is_pressed ? "press" : "release");
+  }
+
+  // Check if this type of sound is muted
+  if (g_mute || mute_state) {
+    if (g_verbose) {
+      printf("Thread %d: %s sound is muted\n", thread_id,
+             is_mouse_event ? "Mouse" : "Keyboard");
+    }
+    goto exit_cleanup;
   }
   if (sound_pack->is_multi) {
     if (is_pressed && sound_pack->multi_key_mappings[key_code].press)
@@ -544,7 +562,8 @@ int parse_keyboard_event(const char *json_line, int *key_code,
 }
 
 void cleanup() {
-  if (g_verbose) printf("Cleaning up...\n");
+  if (g_verbose)
+    printf("Cleaning up...\n");
   usleep(500000);
   for (int i = 0; i < 512; i++) {
     if (g_sound_pack.multi_key_mappings[i].press) {
@@ -569,15 +588,16 @@ static int read_mute_state() {
   if (!rd || strlen(rd) == 0) {
     rd = "/tmp";
   }
-  if (!safe_snprintf(mute_file, sizeof(mute_file), "%s/keyvibe-mute-%d", rd, (int)getuid())) {
+  if (!safe_snprintf(mute_file, sizeof(mute_file), "%s/keyvibe-mute-%d", rd,
+                     (int)getuid())) {
     return 0;
   }
-  
+
   FILE *f = fopen(mute_file, "r");
   if (!f) {
     return 0; // Default to unmuted if file doesn't exist
   }
-  
+
   int mute = 0;
   if (fscanf(f, "%d", &mute) == 1) {
     fclose(f);
@@ -587,15 +607,65 @@ static int read_mute_state() {
   return 0;
 }
 
+static int read_keyboard_mute_state() {
+  char mute_file[MAX_LINE_LENGTH];
+  const char *rd = getenv("XDG_RUNTIME_DIR");
+  if (!rd || strlen(rd) == 0) {
+    rd = "/tmp";
+  }
+  if (!safe_snprintf(mute_file, sizeof(mute_file), "%s/keyvibe-kbd-mute-%d", rd,
+                     (int)getuid())) {
+    return 0;
+  }
+  FILE *f = fopen(mute_file, "r");
+  if (!f) {
+    return 0;
+  }
+  int mute = 0;
+  if (fscanf(f, "%d", &mute) == 1) {
+    fclose(f);
+    return mute;
+  }
+  fclose(f);
+  return 0;
+}
+
+static int read_mouse_mute_state() {
+  char mute_file[MAX_LINE_LENGTH];
+  const char *rd = getenv("XDG_RUNTIME_DIR");
+  if (!rd || strlen(rd) == 0) {
+    rd = "/tmp";
+  }
+  if (!safe_snprintf(mute_file, sizeof(mute_file), "%s/keyvibe-mouse-mute-%d",
+                     rd, (int)getuid())) {
+    return 0;
+  }
+  FILE *f = fopen(mute_file, "r");
+  if (!f) {
+    return 0;
+  }
+  int mute = 0;
+  if (fscanf(f, "%d", &mute) == 1) {
+    fclose(f);
+    return mute;
+  }
+  fclose(f);
+  return 0;
+}
 
 int main(int argc, char *argv[]) {
-  if (argc < 2 || argc > 7) {
-    fprintf(stderr, "Usage: %s <config.json> [volume] [verbose] [mute] [mouse_config] [mouse_volume]\n", argv[0]);
+  if (argc < 2 || argc > 9) {
+    fprintf(stderr,
+            "Usage: %s <config.json> [volume] [verbose] [mute] [mouse_config] "
+            "[mouse_volume] [keyboard_mute] [mouse_mute]\n",
+            argv[0]);
     fprintf(stderr, "  volume: 0-100 (default: 50)\n");
     fprintf(stderr, "  verbose: 1 to enable verbose output (default: 0)\n");
-    fprintf(stderr, "  mute: 1 to mute sound (default: 0)\n");
+    fprintf(stderr, "  mute: 1 to mute all sound (default: 0)\n");
     fprintf(stderr, "  mouse_config: path to mouse config.json (optional)\n");
     fprintf(stderr, "  mouse_volume: 0-100 for mouse volume (optional)\n");
+    fprintf(stderr, "  keyboard_mute: 1 to mute keyboard sounds (optional)\n");
+    fprintf(stderr, "  mouse_mute: 1 to mute mouse sounds (optional)\n");
     return 1;
   }
   if (argc >= 3) {
@@ -605,9 +675,11 @@ int main(int argc, char *argv[]) {
     if (volume_percent > 100)
       volume_percent = 100;
     g_volume = volume_percent / 100.0f;
-    if (g_verbose) printf("Volume set to: %d%%\n", volume_percent);
+    if (g_verbose)
+      printf("Volume set to: %d%%\n", volume_percent);
   } else {
-    if (g_verbose) printf("Volume set to: 50%% (default)\n");
+    if (g_verbose)
+      printf("Volume set to: 50%% (default)\n");
   }
   if (argc >= 4) {
     g_verbose = atoi(argv[3]);
@@ -621,7 +693,7 @@ int main(int argc, char *argv[]) {
       printf("Sound muted\n");
     }
   }
-  
+
   // Parse mouse config and volume if provided
   if (argc >= 6) {
     if (load_sound_config(argv[5]) != 0) {
@@ -634,7 +706,7 @@ int main(int argc, char *argv[]) {
       printf("Mouse sound pack loaded from: %s\n", argv[5]);
     }
   }
-  
+
   if (argc >= 7) {
     int mouse_volume_percent = atoi(argv[6]);
     if (mouse_volume_percent < 0)
@@ -642,9 +714,22 @@ int main(int argc, char *argv[]) {
     if (mouse_volume_percent > 100)
       mouse_volume_percent = 100;
     g_mouse_volume = mouse_volume_percent / 100.0f;
-    if (g_verbose) printf("Mouse volume set to: %d%%\n", mouse_volume_percent);
+    if (g_verbose)
+      printf("Mouse volume set to: %d%%\n", mouse_volume_percent);
   }
-  
+
+  if (argc >= 8) {
+    g_keyboard_mute = atoi(argv[7]);
+    if (g_verbose)
+      printf("Keyboard mute: %s\n", g_keyboard_mute ? "enabled" : "disabled");
+  }
+
+  if (argc >= 9) {
+    g_mouse_mute = atoi(argv[8]);
+    if (g_verbose)
+      printf("Mouse mute: %s\n", g_mouse_mute ? "enabled" : "disabled");
+  }
+
   if (load_sound_config(argv[1]) != 0) {
     fprintf(stderr, "Failed to load keyboard sound configuration\n");
     return 1;
@@ -658,9 +743,11 @@ int main(int argc, char *argv[]) {
   struct timeval timeout;
   char line[MAX_LINE_LENGTH];
   while (1) {
-    // Check mute state from file periodically
+    // Check mute state from files periodically
     g_mute = read_mute_state();
-    
+    g_keyboard_mute = read_keyboard_mute_state();
+    g_mouse_mute = read_mouse_mute_state();
+
     FD_ZERO(&readfds);
     FD_SET(STDIN_FILENO, &readfds);
     timeout.tv_sec = 1;
@@ -682,7 +769,8 @@ int main(int argc, char *argv[]) {
     if (FD_ISSET(STDIN_FILENO, &readfds)) {
       if (fgets(line, sizeof(line), stdin) == NULL) {
         if (feof(stdin)) {
-          if (g_verbose) printf("EOF reached on stdin\n");
+          if (g_verbose)
+            printf("EOF reached on stdin\n");
         } else {
           perror("fgets");
         }
