+++ src/input.c
@@ -1,19 +1,19 @@
+#include "utils.h"
 #include <errno.h>
 #include <stdbool.h>
 #include <stdint.h>
 #include <stdio.h>
-#include "utils.h"
 #include <stdlib.h>
 #include <string.h>
 
 #include <fcntl.h>
 #include <getopt.h>
-#include <poll.h>
-#include <pthread.h>
-#include <unistd.h>
 #include <libevdev/libevdev.h>
 #include <libinput.h>
 #include <libudev.h>
+#include <poll.h>
+#include <pthread.h>
+#include <unistd.h>
 
 #include "config.h"
 
@@ -127,11 +127,13 @@ static int run_mainloop(struct libinput *libinput) {
   fd.events = POLLIN;
   fd.revents = 0;
   if (handle_events(libinput) != 0)
-    return errorf("Expected device added events on startup but got none. Maybe you don't have the right permissions?\n");
+    return errorf("Expected device added events on startup but got none. Maybe "
+                  "you don't have the right permissions?\n");
   while (1) {
     int pr = poll(&fd, 1, -1);
     if (pr < 0) {
-      if (errno == EINTR) continue;
+      if (errno == EINTR)
+        continue;
       return errorf("poll failed: %s\n", strerror(errno));
     }
     handle_events(libinput);
@@ -191,7 +193,8 @@ int main(int argc, char *argv[]) {
   }
   pthread_t input_handler;
   struct input_handler_data input_handler_data = {udev, libinput};
-  if (pthread_create(&input_handler, NULL, handle_input, &input_handler_data) != 0) {
+  if (pthread_create(&input_handler, NULL, handle_input, &input_handler_data) !=
+      0) {
     errorf("Failed to create input handler thread.\n");
   } else {
     pthread_detach(input_handler);
